pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "illustrated-map-generator"
version = "0.1.0"
description = "Generate illustrated maps in the style of Disneyland park maps"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Map Generator"}
]
dependencies = [
    "osmnx>=1.9.0",
    "geopandas>=0.14.0",
    "shapely>=2.0.0",
    "matplotlib>=3.8.0",
    "pillow>=10.0.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "rasterio>=1.3.0",
    "elevation>=1.1.0",
    "google-genai>=1.0.0",
    "pytoshop>=1.2.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]
upscaling = [
    "realesrgan>=0.3.0",
    "basicsr>=1.4.0",
]
psd = [
    "psd-tools>=1.9.0",
]
api = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "pyvips>=2.2.0",
]
all = [
    "realesrgan>=0.3.0",
    "basicsr>=1.4.0",
    "psd-tools>=1.9.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "pyvips>=2.2.0",
]

[project.scripts]
mapgen = "mapgen.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mapgen*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "network: tests that require network access",
    "gemini: tests that call the real Gemini API",
    "slow: tests that take more than 10 seconds",
]
addopts = "-m 'not network and not gemini and not slow'"
